remove source if checked

+++ src/main.py
@@ -80,7 +80,8 @@ class MyImport(sly.app.Import):
                 except Exception as ex:
                     sly.logger.warn(ex)
 
-        remove_dir(context.path)
+        if REMOVE_SOURCE:
+            remove_dir(context.path)
         return context.project_id
